ci: decompose benchmark harness
