+++ .github/workflows/benchmark.yml
@@ -81,31 +81,7 @@ jobs:
 
       - name: Resolve synthetic benchmark checkout
         id: synthetic-benchmark
-        shell: bash
-        run: |
-          set -euo pipefail
-          fixture="benchmarks/synthetic-web-bundler-benchmark"
-          app_root=""
-          tool_root=""
-
-          for checkout in "$GITHUB_WORKSPACE/head" "$GITHUB_WORKSPACE/base"; do
-            if [ -z "$app_root" ] && [ -f "$checkout/$fixture/package.json" ]; then
-              app_root="$checkout/$fixture"
-            fi
-
-            if [ -z "$tool_root" ] && [ -f "$checkout/scripts/bench-synthetic-app.mjs" ]; then
-              tool_root="$checkout"
-            fi
-          done
-
-          if [ -z "$app_root" ] || [ -z "$tool_root" ]; then
-            echo "SKIP=true" >> "$GITHUB_OUTPUT"
-            exit 0
-          fi
-
-          echo "SKIP=false" >> "$GITHUB_OUTPUT"
-          echo "APP_ROOT=$app_root" >> "$GITHUB_OUTPUT"
-          echo "TOOL_ROOT=$tool_root" >> "$GITHUB_OUTPUT"
+        run: node head/scripts/benchmark/ci-resolve-synthetic.mjs
 
       - name: Resolve benchmark-only React plugin
         id: benchmark-imports
@@ -277,21 +253,8 @@ jobs:
           GH_TOKEN: ${{ github.token }}
           PR_NUMBER: ${{ github.event.pull_request.number }}
           REPOSITORY: ${{ github.repository }}
-        run: |
-          set -euo pipefail
-          marker='<!-- react-router-benchmark-ci -->'
-          comment_body="$GITHUB_WORKSPACE/benchmark-output/report/comment.md"
-          body="$(cat "$comment_body")"
-          comment_id="$(gh api "repos/$REPOSITORY/issues/$PR_NUMBER/comments" --paginate \
-            --jq ".[] | select(.body | contains(\"$marker\")) | .id" | tail -n 1)"
-
-          if [ -n "$comment_id" ]; then
-            gh api --method PATCH "repos/$REPOSITORY/issues/comments/$comment_id" \
-              --field body="$body"
-          else
-            gh api --method POST "repos/$REPOSITORY/issues/$PR_NUMBER/comments" \
-              --field body="$body"
-          fi
+          COMMENT_BODY: ${{ github.workspace }}/benchmark-output/report/comment.md
+        run: node head/scripts/benchmark/ci-comment-report.mjs
 
       - name: Persist benchmark history
         if: github.event.pull_request.head.repo.full_name == github.repository
@@ -300,52 +263,4 @@ jobs:
           PR_NUMBER: ${{ github.event.pull_request.number }}
           HEAD_SHA: ${{ github.event.pull_request.head.sha }}
           REPOSITORY: ${{ github.repository }}
-        shell: bash
-        run: |
-          set -euo pipefail
-          history_dir="$GITHUB_WORKSPACE/benchmark-history"
-          remote_url="https://x-access-token:${GH_TOKEN}@github.com/${REPOSITORY}.git"
-
-          set +e
-          git ls-remote --exit-code --heads "$remote_url" "$BENCHMARK_HISTORY_BRANCH" > /dev/null 2>&1
-          branch_status=$?
-          set -e
-
-          if [ "$branch_status" -eq 0 ]; then
-            git clone --depth 1 --branch "$BENCHMARK_HISTORY_BRANCH" "$remote_url" "$history_dir"
-          elif [ "$branch_status" -eq 2 ]; then
-            git clone --depth 1 "$remote_url" "$history_dir"
-            cd "$history_dir"
-            git checkout --orphan "$BENCHMARK_HISTORY_BRANCH"
-            git rm -rf .
-          else
-            exit "$branch_status"
-          fi
-
-          cd "$history_dir"
-          git config user.name "github-actions[bot]"
-          git config user.email "41898282+github-actions[bot]@users.noreply.github.com"
-
-          result_dir="pull-requests/${PR_NUMBER}/${HEAD_SHA}"
-          latest_dir="pull-requests/${PR_NUMBER}/latest"
-          mkdir -p "$result_dir" "$latest_dir"
-          cp "$GITHUB_WORKSPACE/benchmark-output/dev/base/baseline.json" "$result_dir/base.json"
-          cp "$GITHUB_WORKSPACE/benchmark-output/dev/head/baseline.json" "$result_dir/head.json"
-          cp "$GITHUB_WORKSPACE/benchmark-output/build/base/baseline.json" "$result_dir/build-base.json"
-          cp "$GITHUB_WORKSPACE/benchmark-output/build/head/baseline.json" "$result_dir/build-head.json"
-          if [ -d "$GITHUB_WORKSPACE/benchmark-output/synthetic" ]; then
-            cp -R "$GITHUB_WORKSPACE/benchmark-output/synthetic" "$result_dir/synthetic"
-          fi
-          cp "$GITHUB_WORKSPACE/benchmark-output/report/report.json" "$result_dir/report.json"
-          cp "$GITHUB_WORKSPACE/benchmark-output/report/comment.md" "$result_dir/comment.md"
-          cp "$result_dir/report.json" "$latest_dir/report.json"
-          cp "$result_dir/comment.md" "$latest_dir/comment.md"
-
-          git add pull-requests
-          if git diff --cached --quiet; then
-            echo "Benchmark history is already up to date."
-            exit 0
-          fi
-
-          git commit -m "bench: record PR ${PR_NUMBER} ${HEAD_SHA:0:7}"
-          git push origin "$BENCHMARK_HISTORY_BRANCH"
+        run: node head/scripts/benchmark/ci-persist-history.mjs
